+++ Tables.hpp
@@ -4,7 +4,7 @@
 #include "Types.hpp"
 
 extern const int32 g_table[8][4];
-extern const int32 g_table256[8][4];
+extern const int64 g_table256[8][4];
 
 extern const uint32 g_id[4][16];
